# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Parkir",
        "description": "Dokumentasi API Parkir",
        "version": "1.0.0"
    },
    "paths": {
        "/api/parking-blocks": {
            "get": {
                "tags": [
                    "Parking"
                ],
                "summary": "Mengecek ketersediaan blok parkir dan slot",
                "operationId": "getAvailableBlocks",
                "responses": {
                    "200": {
                        "description": "Daftar blok parkir yang tersedia"
                    }
                }
            }
        },
        "/api/parking": {
            "post": {
                "tags": [
                    "Parking"
                ],
                "summary": "Kendaraan parkir di slot parkir pada blok tertentu",
                "operationId": "parkVehicle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "block_id"
                                ],
                                "properties": {
                                    "block_id": {
                                        "description": "ID blok parkir",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kendaraan berhasil parkir"
                    },
                    "422": {
                        "description": "Semua slot sudah terisi"
                    }
                }
            }
        },
        "/api/parking/{blockId}/{slotId}": {
            "delete": {
                "tags": [
                    "Parking"
                ],
                "summary": "Kendaraan keluar dari slot parkir pada blok tertentu",
                "operationId": "exitParking",
                "parameters": [
                    {
                        "name": "blockId",
                        "in": "path",
                        "description": "ID blok parkir",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "slotId",
                        "in": "path",
                        "description": "ID slot parkir",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kendaraan berhasil keluar"
                    },
                    "404": {
                        "description": "Slot parkir tidak tersedia"
                    }
                }
            }
        }
    }
}